requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
tensorflow>=2.13.0
torch>=2.0.0
transformers>=4.30.0

# FPL data fetching
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API framework
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.6.0

# Data processing
joblib>=1.3.0
scipy>=1.11.0

# Sentiment analysis
tweepy>=4.14.0
textblob>=0.17.0
vaderSentiment>=3.3.2

# Model evaluation
matplotlib>=3.7.0
seaborn>=0.12.0

# Scheduling
apscheduler>=3.10.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Environment management
python-dotenv>=1.0.0